requirements.txt
scikit-learn
numpy
python-dotenv
